+++ event.py
@@ -2,11 +2,11 @@ import json
 import datetime
 
 from flask import (
-    Blueprint, flash, g, redirect, render_template, request, url_for, current_app, session
+    Blueprint, flash, g, redirect, render_template, request, url_for, current_app, session, Request, jsonify
 )
 from werkzeug.exceptions import abort
 
-from .auth import login_required
+from .auth import role_required
 
 from .scheduler import scheduler_add_job
 from .utilities.source_utilities import *
@@ -21,6 +21,7 @@ from .config import Config
 bp = Blueprint('event', __name__, url_prefix=Config.URL_PREFIX+'/event')
 
 @bp.route('/source/<sourceId>')
+@role_required("source")
 def source(sourceId):
     allsources = current_app.config['INT2SRC']
     title = allsources[sourceId][0]
@@ -28,6 +29,7 @@ def source(sourceId):
     return render_template('events/source-events.html', allsources=allsources, sourceId=sourceId, title=title, calendars=calendars, total=0, eventTypeValues=eventTypeValues)
 
 @bp.route('/calendar/<calendarId>')
+@role_required("source")
 def calendar(calendarId):
     if 'select_status' in session:
         select_status = session['select_status']
@@ -65,18 +67,29 @@ def calendar(calendarId):
                             pagination=pagination, eventTypeValues=eventTypeValues)
 
 @bp.route('/setting', methods=('GET', 'POST'))
-@login_required
+@role_required("source")
 def setting():
     if request.method == 'POST':
         print(request.form)
+        #add update calendars
         allstatus = get_all_calendar_status()
         update_calendars_status(request.form, allstatus)
-    allstatus = get_all_calendar_status()
-    return render_template('events/setting.html', sources=current_app.config['INT2SRC'], allstatus=allstatus)
-
+    calendar_in_db = get_all_calendar_status()
+    calendar_ids = calendar_in_db.keys()
+    calendar_source = list()
+    calendar_status = dict()
+    for calendar_id in calendar_ids:
+        calendar_source.append({calendar_id: calendar_in_db.get(calendar_id).get('calendarName')})
+        calendar_status[calendar_id] = calendar_in_db.get(calendar_id).get('status')
+    INT2SRC = {
+        '0': ('WebTools', calendar_source),
+        '1': ('EMS', []),
+    }
+    calendar_prefix=current_app.config['WEBTOOL_CALENDAR_LINK_PREFIX']
+    return render_template('events/setting.html', sources=INT2SRC, allstatus=calendar_status, url_prefix=calendar_prefix)
 
 @bp.route('/download', methods=['POST'])
-@login_required
+@role_required("source")
 def download():
     targets = request.get_json()
     if targets:
@@ -84,7 +97,7 @@ def download():
     return json.dumps({'status': 'OK', 'data': 'complete'})
 
 @bp.route('/select', methods=['POST'])
-@login_required
+@role_required("source")
 def select():
     select_status = []
     if request.form.get('approved') == '1':
@@ -100,32 +113,33 @@ def select():
     return "", 200
 
 @bp.route('/approve', methods=('GET', 'POST'))
-@login_required
+@role_required("source")
 def approveCalendar():
     calendarId = request.form['calendarId']
     approve_calendar_db(calendarId)
     return "success", 200
 
 @bp.route('/disapprove', methods=('GET', 'POST'))
-@login_required
+@role_required("source")
 def disapproveCalendar():
     calendarId = request.form['calendarId']
     disapprove_calendar_db(calendarId)
     return "success", 200
 
 @bp.route("/approveEvent/<id>", methods=['GET', 'POST'])
-@login_required
+@role_required("source")
 def approveEvent(id):
     approve_event(id)
     return "success", 200
 
 @bp.route("/disapproveEvent/<id>", methods=['GET', 'POST'])
-@login_required
+@role_required("source")
 def disapproveEvent(id):
     disapprove_event(id)
     return "success", 200
 
 @bp.route('/detail/<eventId>')
+@role_required("source")
 def detail(eventId):
     event = get_event(eventId)
     source = current_app.config['INT2SRC'][event['sourceId']]
@@ -139,6 +153,7 @@ def detail(eventId):
 
 
 @bp.route('/edit/<eventId>', methods=('GET', 'POST'))
+@role_required("source")
 def edit(eventId):
     post_by_id = get_event(eventId)
     if request.method == 'POST':
@@ -162,6 +177,7 @@ def edit(eventId):
     return render_template("events/event-edit.html", post = post_by_id, eventTypeMap = eventTypeMap, eventTypeValues=eventTypeValues, isUser=False, sourceName=sourceName, calendarName=calendarName)
 
 @bp.route('/searchresult', methods=['GET'])
+@role_required("source")
 def searchresult():
     if 'select_status' in session:
         select_status = session['select_status']
@@ -199,6 +215,7 @@ def searchresult():
     )
 
 @bp.route('/schedule', methods=['POST'])
+@role_required("source")
 def schedule():
     time = request.form.get('time')
     targets = json.loads(request.form.get('targets'))
@@ -213,3 +230,36 @@ def schedule():
     print(time)
     scheduler_add_job(current_app._get_current_object(), current_app.scheduler, start, time, targets=targets)
     return "success", 200
+
+
+@bp.route('/add-new-calendar', methods=['POST'])
+@role_required("source")
+def add_new_calendar():
+    print(request.form)
+    # new calendars
+    calendarID = request.form.get('data[calendarID]')
+    calendarName = request.form.get('data[calendarName]')
+    print(calendarID)
+    print(calendarName)
+    if calendarID == '' or calendarName == '':
+        print("should have both ID and Name!")
+        return "invalid", 200
+    # all newly added calendar will be default to "disapproved"
+    calendar_document = {"calendarId" : calendarID, "calendarName": calendarName, "status": "disapproved"}
+    insert_result = insert_one(current_app.config['CALENDAR_COLLECTION'], document = calendar_document)
+    # insert error condition check
+    if insert_result.inserted_id is None:
+        print("Insert calendar " + calendarID +" failed")
+        return redirect('event.setting')
+        return "fail", 400
+    else:
+        print(current_app.config['INT2CAL'])
+        print("successfully inserted calendar "+ calendarID)
+        return "success", 200
+
+@bp.route('/searchtip', methods=["POST"]) 
+@role_required("source")
+def searchtip():
+    data = Request.get_json()
+    print(data)
+    return jsonify([{ "title": "test1"}, { "title": "test2"}, { "title": "test3"}])